Lexer: do not split lines after segment override (A:, PP:, XP:, YP:, ZP:)

+++ src/z80asm/z80asm2/file_cache.cpp
@@ -448,7 +448,7 @@ const std::vector<std::string>& DependencyFile::dependencies() const {
 }
 
 bool DependencyFile::write() {
-    const unsigned LINE_WIDTH = 80;
+    const size_t LINE_WIDTH = 80;
 
     if (dependencies_.empty()) {
         return false;
